+++ regression-test/data/fault_injection_p0/fault_injection_p0/test_segcompaction_fault_injection.out
@@ -1,0 +1,7 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !select_default --
+
+-- !select_default --
+
+-- !select_default --
+
