+++ regression-test/data/fault_injection_p0/fault_injection_p0/test_too_many_segments_fault_injection.out
@@ -1,0 +1,3 @@
+-- This file is automatically generated. You should know what you did if you want to edit this
+-- !select_default --
+
