+++ regression-test/suites/fault_injection_p0/test_segcompaction_fault_injection.groovy
@@ -141,88 +141,3 @@ suite("test_segcompaction_correctness") {
     }
 }
 
-suite("test_segcompaction_too_many_segments") { // the epic -238 case
-    def runLoadWithTooManySegments = {
-        String ak = getS3AK()
-        String sk = getS3SK()
-        String endpoint = getS3Endpoint()
-        String region = getS3Region()
-        String bucket = getS3BucketName()
-        def backendId_to_backendIP = [:]
-        def backendId_to_backendHttpPort = [:]
-        String backend_id;
-        try {
-            getBackendIpHttpPort(backendId_to_backendIP, backendId_to_backendHttpPort);
-            backend_id = backendId_to_backendIP.keySet()[0]
-            def (code, out, err) = show_be_config(backendId_to_backendIP.get(backend_id), backendId_to_backendHttpPort.get(backend_id))
-
-            logger.info("Show config: code=" + code + ", out=" + out + ", err=" + err)
-            assertEquals(code, 0)
-            def configList = parseJson(out.trim())
-            assert configList instanceof List
-
-            boolean disableAutoCompaction = true
-            for (Object ele in (List) configList) {
-                assert ele instanceof List<String>
-                if (((List<String>) ele)[0] == "disable_auto_compaction") {
-                    disableAutoCompaction = Boolean.parseBoolean(((List<String>) ele)[2])
-                }
-            }
-            sql """ DROP TABLE IF EXISTS ${tableName} """
-            sql "${create_table_sql}"
-
-            def uuid = UUID.randomUUID().toString().replace("-", "0")
-            String columns_str = ("$columns" != "") ? "($columns)" : "";
-
-            sql """
-            LOAD LABEL $uuid (
-                DATA INFILE("s3://$bucket/regression/segcompaction/segcompaction.orc")
-                INTO TABLE ${tableName}
-                FORMAT AS "ORC"
-                $columns_str
-            )
-            WITH S3 (
-                "AWS_ACCESS_KEY" = "$ak",
-                "AWS_SECRET_KEY" = "$sk",
-                "AWS_ENDPOINT" = "$endpoint",
-                "AWS_REGION" = "$region"
-            )
-            properties(
-                "use_new_load_scan_node" = "true"
-            )
-            """
-
-            Thread.sleep(2000)
-            GetDebugPoint().enableDebugPointForAllBEs("BetaRowsetWriter._check_segment_number_limit_too_many_segments",
-                                                      ["segnum":2000])
-
-            def max_try_milli_secs = 3600000
-            String [][] result = ''
-            while (max_try_milli_secs > 0) {
-                result = sql """ show load where label="$uuid" order by createtime desc limit 1; """
-                if (result[0][2].equals("FINISHED")) {
-                    logger.info("Load FINISHED " + " $uuid")
-                    break;
-                }
-                if (result[0][2].equals("CANCELLED")) {
-                    logger.info("Load CANCELLED " + " $uuid")
-                    break;
-                }
-                Thread.sleep(1000)
-                max_try_milli_secs -= 1000
-                if(max_try_milli_secs <= 0) {
-                    assertTrue(1 == 2, "load Timeout: $uuid")
-                }
-            }
-            assertTrue(result[0][7].contains("-238")) // EPIC!
-
-            result = sql """ show load where label="$uuid" order by createtime desc limit 1; """
-            qt_select_default """ SELECT * FROM ${tableName} WHERE col_0=47 order by col_1, col_2; """
-            tablets = sql """ show tablets from ${tableName}; """
-        } finally {
-            try_sql("DROP TABLE IF EXISTS ${tableName}")
-            GetDebugPoint().disableDebugPointForAllBEs("BetaRowsetWriter._check_segment_number_limit_too_many_segments")
-        }
-    }
-    runLoadWithTooManySegments()
-}
